Fix starting fun with non-binding
Before, when starting with a binding, the first term
was an all-inclusive binding.  So the function body was one
term by coinsidence.

If you had a function call as the first term in the function, you
ended up with multiple arguments to the function creating macro which
bailed out.  So, wrap all local fun bodies in a progn so every
local function body ends up being one term.