Force matching of ([blah] : xs) when blah is defined
This isn't completely pretty and it violates the "everything
matches as a list containing lists" constraint AND it throws
away any (: xs) arg just because it doesn't make sense to recurse
if the first match didn't happen anyway (you can't get there).

So, the (: xs) signifies "we have a rest, but we won't have
access to it."  Should probably just use (field : _) instead.